pyproject.toml
[project]
name = "convo"
version = "0.2.0"
description = "Save conversations as clean, formatted documents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click",
    "pyyaml",
    "reportlab",
    "jinja2",
]

[project.optional-dependencies]
private = [
    "presidio-analyzer",
    "presidio-anonymizer",
    "spacy",
]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
convo = "convo.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["convo"]

[tool.ruff]
target-version = "py313"
line-length = 100
